requirements/base.txt
Django==2.0.0
django-environ
djangorestframework
